requirements.txt
build==1.2.2.post1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorclass==2.2.2
confluent-kafka==2.10.0
docopt==0.6.2
hvac==2.3.0
idna==3.10
Naked==0.1.32
packaging==25.0
pip-tools==7.4.1
pip-upgrader==1.4.15
pycryptodome==3.23.0
pyhcl==0.4.5
pyproject_hooks==1.2.0
PyYAML==6.0.2
requests==2.32.3
setuptools==80.8.0
shellescape==3.8.1
terminaltables==3.1.10
urllib3==2.4.0
wheel==0.45.1
